switch to double precision almost everywhere some bugfixes common rand function

+++ include/mingafactoryparams.h
@@ -13,7 +13,7 @@ public:
 	bool write(serut::SerializationInterface &si) const;
 	bool read(serut::SerializationInterface &si);
 
-	float *dij;
+	double *dij;
 	int *minDose;
 	int *maxDose;
 };
